+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2.1.2
+
+- Update dependencies
+
 ## 2.1.1
 
 - Fix prerelease dependency